requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
requests>=2.31.0
pytrends>=4.9.0
fredapi>=0.5.0
vaderSentiment>=3.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlalchemy>=2.0.0
